// src/http/generated/api.ts
export type GetWeekSummary200SummaryGoalsPerDay = { [key: string]: unknown } | null;

export type GetWeekSummary200Summary = {
  completed: number;
  /** @nullable */
  total: number | null;
  /** @nullable */
  goalsPerDay: GetWeekSummary200SummaryGoalsPerDay;
};